Refactor addons into multiple packages
