+++ cmd/kubeadm/app/phases/addons/proxy/BUILD
@@ -6,11 +6,19 @@ load(
     "go_test",
 )
 
+go_test(
+    name = "go_default_test",
+    srcs = ["proxy_test.go"],
+    library = ":go_default_library",
+    tags = ["automanaged"],
+    deps = ["//cmd/kubeadm/app/util:go_default_library"],
+)
+
 go_library(
     name = "go_default_library",
     srcs = [
-        "addons.go",
         "manifests.go",
+        "proxy.go",
     ],
     deps = [
         "//cmd/kubeadm/app/apis/kubeadm:go_default_library",
@@ -18,9 +26,11 @@ go_library(
         "//cmd/kubeadm/app/util:go_default_library",
         "//cmd/kubeadm/app/util/apiclient:go_default_library",
         "//pkg/api:go_default_library",
+        "//pkg/util/version:go_default_library",
         "//plugin/pkg/scheduler/algorithm:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/api/extensions/v1beta1:go_default_library",
+        "//vendor/k8s.io/api/rbac/v1beta1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
@@ -39,11 +49,4 @@ filegroup(
     name = "all-srcs",
     srcs = [":package-srcs"],
     tags = ["automanaged"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["addons_test.go"],
-    library = ":go_default_library",
-    deps = ["//cmd/kubeadm/app/util:go_default_library"],
 )
